requirements.txt
Django>=5.0.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers